Partitioned Least Squares

Linear least squares is one of the most widely used regression methods among scientists in many fields. The simplicity of the model allows this method to be used when data is scarce and it is usually appealing to practitioners that need to gather some insight into the problem by inspecting the values of the learnt parameters. PartitionedLS is a variant of the linear least squares model allowing practitioners to partition the input features into groups of variables that they require to contribute similarly to the final result.

The model

The Partitioned Least Squares model is formally defined as:

\[\begin{gather*} \text{minimize}_{\mathbf{\alpha}, \mathbf{\beta}} \| \mathbf{X} \times (\mathbf{P} \circ \mathbf{\alpha}) \times \mathbf{\beta} - \mathbf{y} \|_2^2 \\ \begin{aligned} \quad s.t.\quad &\mathbf{\alpha} \succeq 0\\ &\mathbf{P}^T \times \mathbf{\alpha} = \mathbf{1}. \end{aligned} \end{gather*}\]

where:

$\mathbf{X}$ is $N \times M$ data matrix;

$\mathbf{P}$ is a user-defined partition matrix having $K$ columns (one for each element of the partition), $M$ rows, and containing $1$ in $P_{i,j}$ if the $i$-th attribute belongs to the $j$-th partition and $0$ otherwise;

$\mathbf{\beta}$ is a vector weighting the importance of each set of attributes in the partition;

$\mathbf{\alpha}$ is a vector weighting the importance of each attribute within one of the sets in the partition. Note that the constraints imply that for each set in the partition the weights of the corresponding $\alpha$ variables are all positive and sum to $1$.

The PartitionedLS problem is non-convex and NP-complete. The library provides two algorithms to solve the problem anyway: an iterative algorithm based on the Alternating Least Squares approach and an optimal algorithm that guarantees requiring however exponential time in the cardinality of the partition (i.e., it is mainly useful when $K$ is small).

More details can be found in the paper Partitioned Least Squares.

To install this library

Just add it as a dependency to your Julia environment. Launch julia from the main directory of your project and enter the following commands:

# Opens the package manager REPL
]

# Activate you local environment (can be skipped if you want to install the library globally)
activate .

# Adds the library to the environment
add PartitionedLS

To use this library

You will need a matrix P describing the partitioning of your variables, e.g.:

P = [[1 0]; 
     [1 0]; 
     [0 1]]

specifies that the first and the second variable belongs to the first partition, while the third variable belongs to the second.

You then just give your data to the fit function and use the predict function to make predictions.

A complete example:

using PartitionedLS

X = [[1. 2. 3.]; 
     [3. 3. 4.]; 
     [8. 1. 3.]; 
     [5. 3. 1.]]

y = [1.; 
     1.; 
     2.; 
     3.]

P = [[1 0]; 
     [1 0]; 
     [0 1]]

# fit using the optimal algorithm 
result_opt = fit(Opt, X, y, P, η = 0.0)

# fit using the iterative algorithm
result_alt = fit(Alt, X, y, P, η = 0.0)

# fit using the BnB algorithm
result_alt_nnls = fit(BnB, X, y, P)

# Make predictions on the given data matrix. The function works
# with results returned by anyone of the solvers.
predict(result_opt.model, X)

Function Documentation

PartitionedLS.fit — Function

fit(::Type{Alt}, X::Matrix{Float64}, y::Vector{Float64}, P::Matrix{Int}; η = 0.0, ϵ = 1e-6, T = 100, nnlsalg = :nnls)

Fits a PartitionedLS model by alternating the optimization of the α and β variables. This version uses an optimization strategy based on non-negative-least-squaes solvers. This formulation is faster and more numerically stable with respect to fit(Alt, ...)`.

Arguments

X: $N × M$ matrix describing the examples

y: $N$ vector with the output values for each example

P: $M × K$ matrix specifying how to partition the $M$ attributes into $K$ subsets. $P_{m,k}$ should be 1 if attribute number $m$ belongs to

partition $k$.

η: regularization factor, higher values implies more regularized solutions

T: number of alternating loops to be performed, defaults to 1000.

ϵ: minimum relative improvement in the objective function before stopping the optimization. Default is 1e-6

nnlsalg: specific flavour of nnls algorithm to be used, possible values are :pivot, :nnls, :fnnls, default is :nnls

Result

A NamedTuple with the following fields:

opt: optimal value of the objective function (loss + regularization)

model: a NamedTuple containing the following fields:

a: values of the α variables at the optimal point

b: values of the β variables at the optimal point

t: the intercept at the optimal point

P: the partition matrix (copied from the input)

fit(::Type{Opt}, X::Matrix{Float64}, y::Vector{Float64}, P::Matrix{Int}; η=0.0, returnAllSolutions=false, nnlsalg=:nnls)

Fits a PartialLS Regression model to the given data and resturns the learnt model (see the Result section). It uses a coplete enumeration strategy which is exponential in K, but guarantees to find the optimal solution.

Arguments

X: $N × M$ matrix describing the examples

y: $N$ vector with the output values for each example

P: $M × K$ matrix specifying how to partition the $M$ attributes into $K$ subsets. $P_{m,k}$ should be 1 if attribute number $m$ belongs to

partition $k$.

η: regularization factor, higher values implies more regularized solutions (default: 0.0)

returnAllSolutions: if true an additional output is appended to the resulting tuple containing all solutions found during the algorithm.

nnlsalg: the kind of nnls algorithm to be used during solving. Possible values are :pivot, :nnls, :fnnls (default: :nnls)

Result

A NamedTuple with the following fields:

opt: optimal value of the objective function (loss + regularization)

model: a NamedTuple containing the following fields:

a: values of the α variables at the optimal point

b: values of the β variables at the optimal point

t: the intercept at the optimal point

P: the partition matrix (copied from the input)

solutions: all solutions found during the execution (returned only if resultAllSolutions=true)

Example

X = rand(100, 10)
y = rand(100)
P = [1 0 0; 0 1 0; 0 0 1; 1 1 0; 0 1 1]
result = fit(Opt, X, y, P)

fit(::Type{BnB}, X::Matrix{Float64}, y::Vector{Float64}, P::Matrix{Int}, η=0.0, nnlsalg=:nnls)

Implements the Branch and Bound algorithm to fit a Partitioned Least Squres model.

Arguments

X: $N × M$ matrix describing the examples

y: $N$ vector with the output values for each example

P: $M × K$ matrix specifying how to partition the $M$ attributes into $K$ subsets. $P_{m,k}$ should be 1 if attribute number $m$ belongs to

partition $k$.

η: regularization factor, higher values implies more regularized solutions (default: 0.0)

nnlsalg: the kind of nnls algorithm to be used during solving. Possible values are :pivot, :nnls, :fnnls (default: :nnls)

Result

A tuple of the form: (opt, a, b, t, P, nopen)

opt: optimal value of the objective function (loss + regularization)

model: a NamedTuple containing the following fields:

a: values of the α variables at the optimal point

b: values of the β variables at the optimal point

t: the intercept at the optimal point

P: the partition matrix (copied from the input)

nopen: the number of nodes opened by the BnB algorithm

PartitionedLS.predict — Function

predict(α::Vector{Float64}, β::Vector{Float64}, t::Float64, P::Matrix{Int}, X::Matrix{Float64})::Vector{Float64}

Result

the prediction for the partitioned least squares problem with solution α, β, t over the dataset X and partition matrix P

predict(model, X)::Vector{Float64}

Make predictions for the datataset X using the PartialLS model model.

Arguments

model is a NamedTuple in the form returned by fit functions, it shall contains the following elements:

opt: the optimal value of the objective attained by the fit function

'model': a NamedTuple containing the following elements:

α: the values of the α variables

β: the values of the β variables

t: the value of the t variable

P: the partition matrix

X: the data containing the examples for which the predictions are sought

Return

the predictions of the given model on examples in X.