PartitionedLS.jl

TBD

Package Features

TBD

Function Documentation

PartitionedLS.fit — Function

fit(::Type{Alt}, X::Array{Float64,2}, y::Array{Float64,1}, P::Array{Int,2}; beta=randomvalues)

Fits a PartitionedLS model by alternating the optimization of the α and β variables.

Arguments

X: $N × M$ matrix describing the examples

y: $N$ vector with the output values for each example

P: $M × K$ matrix specifying how to partition the $M$ attributes into $K$ subsets. $P_{m,k}$ should be 1 if attribute number $m$ belongs to

partition $k$.

η: regularization factor, higher values implies more regularized solutions

T: number of alternating loops to be performed, defaults to 20.

get_solver: a function returning the solver to be used. Defaults to () -> ECOSSolver()

resume and checkpoint allows for restarting an optimization from a given checkpoint.

Result

A tuple of the form: (opt, a, b, t, P)

opt: optimal value of the objective function (loss + regularization)

a: values of the α variables at the optimal point

b: values of the β variables at the optimal point

t: the intercept at the optimal point

P: the partition matrix (copied from the input)

fit(::Type{Opt}, X::Array{Float64,2}, y::Array{Float64,1}, P::Array{Int,2}; beta=randomvalues)

Fits a PartialLS Regression model to the given data and resturns the learnt model (see the Result section).

Arguments

X: $N × M$ matrix describing the examples

y: $N$ vector with the output values for each example

P: $M × K$ matrix specifying how to partition the $M$ attributes into $K$ subsets. $P_{m,k}$ should be 1 if attribute number $m$ belongs to

partition $k$.

η: regularization factor, higher values implies more regularized solutions

get_solver: a function returning the solver to be used. Defaults to () -> ECOSSolver()

Result

A tuple of the form: (opt, a, b, t, P)

opt: optimal value of the objective function (loss + regularization)

a: values of the α variables at the optimal point

b: values of the β variables at the optimal point

t: the intercept at the optimal point

P: the partition matrix (copied from the input)

The output model predicts points using the formula: f(X) = $X * (P .* a) * b + t$.

fit(::Type{BnB}, X::Array{Float64,2}, y::Array{Float64,1}, P::Array{Int,2}, η=1.0, nnlsalg=:pivot)

Implements the Branch and Bound algorithm to fit a Partitioned Least Squres model.

Arguments

X: $N × M$ matrix describing the examples

y: $N$ vector with the output values for each example

P: $M × K$ matrix specifying how to partition the $M$ attributes into $K$ subsets. $P_{m,k}$ should be 1 if attribute number $m$ belongs to

partition $k$.

η: regularization factor, higher values implies more regularized solutions

nnlsalg: the kind of nnls algorithm to be used during solving

Result

A tuple of the form: (opt, a, b, t, P, nopen)

opt: optimal value of the objective function (loss + regularization)

a: values of the α variables at the optimal point

b: values of the β variables at the optimal point

t: the intercept at the optimal point

P: the partition matrix (copied from the input)

nopen: the number of nodes opened by the BnB algorithm

The output model predicts points using the formula: f(X) = $X * (P .* a) * b + t$.

PartitionedLS.predict — Function

predict(model::Tuple, X::Array{Float64,2})

returns the predictions of the given model on examples in X.